We are a collaborative financing factoring platform, aimed at SMEs throughout Chile.  Mundialis is a fintech company focused on holding online meetings between investors and clients who require financing to continue the development and growth of their companies. We are committed to the development of the digital financial industry, to have a more fair and transparent society, improving financing opportunities for small and medium-sized companies at fair rates, determined by the market.
We are a collaborative financing factoring platform, aimed at SMEs throughout Chile.

Mundialis is a fintech company focused on holding online meetings between investors and clients who require financing to continue the development and growth of their companies. We are committed to the development of the digital financial industry, to have a more fair and transparent society, improving financing opportunities for small and medium-sized companies at fair rates, determined by the market.